+++ sbbw/src/main.rs
@@ -1,68 +1,34 @@
 #![feature(proc_macro_hygiene, decl_macro, str_split_as_str)]
 use actix_web::{App, HttpServer};
-use cmd::get_args;
+use cmd::{
+    args::{get_args, to_request, WidgetCommands},
+    install::install_widget,
+};
 use colored::*;
-use fork::fork;
+use fork::{fork, Fork};
 use sbbw_exec::autostarts;
-use sbbw_widget_conf::{get_widgets, get_widgets_path, validate_config_toml};
+use sbbw_widget_conf::{
+    exits_widget, generate_config_sbbw, generate_pid_file, get_config_sbbw, get_widgets,
+    get_widgets_path, validate_config_toml, SbbwConfig, get_pid,
+};
 use std::{env, process::Command};
 use widget::routes;
 
 mod cmd;
 mod widget;
 
-const VERSION: &str = env!("CARGO_PKG_VERSION");
 const DESCRIPTION: &str = env!("CARGO_PKG_DESCRIPTION");
 const AUTHORS: &str = env!("CARGO_PKG_AUTHORS");
 
-#[actix_web::main]
-async fn main() -> Result<(), std::io::Error> {
+fn main() {
     let widgets = get_widgets();
     let args = get_args().lock().unwrap();
 
     if args.show_windows {
         println!("{}", "Installed widgets:".blue().bold());
-        for widget in widgets {
+        for widget in &widgets {
             println!("{} {}", "-".blue(), widget);
         }
-        return Ok(());
-    }
-
-    println!("{}", "Sbbw Daemon".green());
-
-    if let Some(value) = &args.check_config {
-        if widgets.contains(&value.clone()) {
-            let path_conf = get_widgets_path().join(value).join("config.toml");
-            if path_conf.exists() {
-                match validate_config_toml(path_conf) {
-                    Ok(_) => {
-                        println!(
-                            "{}",
-                            "Config of widget {} is valid"
-                                .green()
-                                .replace("{}", &value.yellow().bold())
-                        );
-                    }
-                    Err(e) => {
-                        println!("{e}");
-                        println!(
-                            "{}",
-                            "Config of widget {} is not valid"
-                                .red()
-                                .replace("{}", &value.yellow().bold())
-                        );
-                    }
-                }
-                return Ok(());
-            }
-        }
-        println!(
-            "{}",
-            "Widget {} not found"
-                .red()
-                .replace("{}", &value.yellow().bold())
-        );
-        return Ok(());
     }
 
     match Command::new("sbbw-widget").spawn() {
@@ -76,20 +42,109 @@ async fn main() -> Result<(), std::io::Error> {
                 "Binary for launch Widgets not found".red().bold(),
                 e
             );
-            return Ok(());
         }
     }
 
     autostarts();
 
-    if !args.no_fork {
-        fork().unwrap();
-    }
+    if let Some(cmd) = &args.widget_cmd {
+        let port = match get_config_sbbw() {
+            Ok(cfg) => cfg.port,
+            Err(_) => args.port,
+        };
 
-    HttpServer::new(move || App::new().configure(routes))
-        .bind(("0.0.0.0", args.port))
-        .unwrap()
-        .workers(4)
-        .run()
-        .await
+        match cmd {
+            WidgetCommands::Run => {
+                if get_pid().is_err() {
+                    println!("{}", "Other Sbbw Daemon is running".red().bold());
+                    std::process::exit(1);
+                }
+                println!("{}", "Sbbw Daemon".green());
+                if !args.no_fork {
+                    match fork() {
+                        Ok(Fork::Parent(child)) => {
+                            generate_pid_file(child.to_string());
+                            std::process::exit(0);
+                        }
+                        _ => println!("Cannot create fork"),
+                    }
+                } else {
+                    generate_pid_file(std::process::id().to_string());
+                }
+                generate_config_sbbw(SbbwConfig {
+                    port: args.port,
+                    ..Default::default()
+                })
+                .unwrap();
+
+                let server = HttpServer::new(move || App::new().configure(routes))
+                    .bind(("0.0.0.0", port))
+                    .unwrap()
+                    .workers(4)
+                    .run();
+                match actix_rt::System::new().block_on(server) {
+                    Ok(_) => println!("Done"), // Call when the server is closed
+                    Err(e) => println!("Error: {}", e),
+                }
+            }
+            WidgetCommands::Check { widget_name } => {
+                if exits_widget(widget_name.to_string()) {
+                    let path_conf = get_widgets_path().join(widget_name).join("config.toml");
+                    match validate_config_toml(path_conf) {
+                        Ok(_) => {
+                            println!(
+                                "{}",
+                                "Config of widget {} is valid"
+                                    .green()
+                                    .replace("{}", &widget_name.yellow().bold())
+                            );
+                        }
+                        Err(e) => {
+                            println!("{e}");
+                            println!(
+                                "{}",
+                                "Config of widget {} is not valid"
+                                    .red()
+                                    .replace("{}", &widget_name.yellow().bold())
+                            );
+                        }
+                    }
+                }
+            }
+            WidgetCommands::Install {
+                repo,
+                new_name,
+                service,
+            } => {
+                match install_widget(WidgetCommands::Install {
+                    repo: repo.to_string(),
+                    new_name: new_name.to_owned(),
+                    service: service.to_owned(),
+                }) {
+                    Err(e) => println!("{e}"),
+                    Ok(_) => println!("Success Widget installed"),
+                }
+            }
+            x => match to_request(x, format!("http://localhost:{}", port)) {
+                Ok(req) => {
+                    let client = reqwest::Client::new();
+                    let post = client
+                        .post(format!("http://localhost:{}/rpc", port))
+                        .json(&req)
+                        .send();
+                    match actix_rt::System::new().block_on(post) {
+                        Ok(res) => {
+                            if !res.status().is_success() {
+                                println!("Cannot comunicate with daemon");
+                            }
+                        },
+                        Err(e) => println!("Error calling daemon: {}", e),
+                    }
+                }
+                Err(e) => {
+                    println!("Failed map request: {e}")
+                }
+            },
+        };
+    }
 }